+++ src/DICTrait.php
@@ -38,6 +38,7 @@ trait DICTrait {
 	 * @return PluginInterface Plugin interface
 	 *
 	 * @throws DICException Class $plugin_class_name not exists!
+	 * @throws DICException Class $plugin_class_name not extends ilPlugin!
 	 * @logs   DEBUG Please implement $plugin_class_name::getInstance()!
 	 */
 	protected static final function plugin() {